add OrderItem implements many-to-may association with extra attributes and update type price to BigDecimal

+++ src/main/java/com/compassoul/workshop/entities/Product.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 
 import java.io.Serial;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
@@ -19,7 +20,7 @@ public class Product implements Serializable {
     private Long id;
     private String name;
     private String description;
-    private Double price;
+    private BigDecimal price;
     private String imageUrl;
 
     @ManyToMany
@@ -32,7 +33,7 @@ public class Product implements Serializable {
         super();
     }
 
-    public Product(Long id, String name, String description, Double price, String imageUrl) {
+    public Product(Long id, String name, String description, BigDecimal price, String imageUrl) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -64,11 +65,11 @@ public class Product implements Serializable {
         this.description = description;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
